Build default image from a Containerfile
Switch to debian buster as the base image from ubi8. Do multi-stage
build using a rust builder image (which is based on debian buster).

Why ?
Solve the issue of having a newer glibc during runtime - when building
binary on fedora and packaging into ubi8. The base is debian buster
for both build-time and runtime now.